+++ test/unit/bindings/test_functions.py
@@ -9,9 +9,10 @@ import json
 
 from functools import partial
 from collections import namedtuple
-from unittest.mock import patch, MagicMock
+from unittest.mock import patch, MagicMock, Mock
 
-from pureport import api
+from pureport.bindings import functions
+from pureport.session import Session
 
 SentinelResponse = namedtuple('SentinelResponse', ('status', 'data', 'headers', 'json'))
 response = partial(SentinelResponse, status=200, data=None, headers=None, json=None)
@@ -20,21 +21,26 @@ response = partial(SentinelResponse, status=200, data=None, headers=None, json=N
 def generate_response(method, url, body=None, headers=None, query=None):
     if url == '/openapi.json':
         basepath = os.path.dirname(__file__)
-        content = open(os.path.join(basepath, '../openapi.json')).read()
+        content = open(os.path.join(basepath, '../../openapi.json')).read()
         return response(data=content, json=json.loads(content))
 
     elif url == '/accounts':
         return response(data=json.dumps([]), json=[])
 
 
-@patch.object(api, 'Session')
-def test_api_methods_exist(mock_session):
+@patch.object(functions, 'get_api')
+def test_functions_methods_exist(mock_get_api):
+    session = Mock(spec=Session)
+
     basepath = os.path.dirname(__file__)
-    content = open(os.path.join(basepath, '../openapi.json')).read()
-    mock_session.get.return_value = content
+    content = open(os.path.join(basepath, '../../openapi.json')).read()
+
+    mock_get_api.return_value = json.loads(content)
+
+    functions.make(session)
 
     methods = re.findall('operationId: (.+)', content, re.M)
-    assert set(methods).issubset(dir(api))
+    assert set(methods).issubset(dir(session))
 
 
 def test_request():
@@ -43,16 +49,4 @@ def test_request():
     mock_session.side_effect = partial(generate_response)
     mock_session.get.side_effect = partial(generate_response, 'GET')
 
-    api.request(mock_session, 'GET', '/accounts')
-
-
-@patch.object(api, 'get_api')
-def test_make(mock_get_api):
-    basepath = os.path.dirname(__file__)
-    content = open(os.path.join(basepath, '../openapi.json')).read()
-    mock_get_api.return_value = json.loads(content)
-
-    api.make()
-
-    methods = re.findall('operationId: (.+)', content, re.M)
-    assert set(methods).issubset(dir(api))
+    functions.request(mock_session, 'GET', '/accounts')