+++ CHANGELOG.md
@@ -4,6 +4,11 @@
   * Added describe() function to recreate entire model schema
   * Fixed documentation formating issue (#16)
   * Remove unused setting `pureport.defaults.account_id`
+  * Refactor bindings system to make functions session-aware
+  * Automatically determine account ID for a session
+  * Automatically inject session account id into function arguments
+  * Remove `pureport.defaults.automake_bindings` setting
+  * Add automake_bindings keyword argument to session init
  
 
 0.2.0 / 2020-09-07